+++ BusRouteLondon.Web/Models/BusStop.cs
@@ -4,10 +4,10 @@ namespace BusRouteLondon.Web
     {
         public string BusStopCode { get; set; }
         public string BusStopName { get; set; }
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
         
-        public decimal Easting { get; set; }
-        public decimal Northing { get; set; }
+        public int Easting { get; set; }
+        public int Northing { get; set; }
     }
 }
